Change variable display types

+++ BitMagic.X16Debugger/StackManager.cs
@@ -68,7 +68,7 @@ internal class StackManager
             _data.Add(new Variable()
             {
                 Name = i.ToString(),
-                Type = "int",
+                Type = "byte",
                 PresentationHint = new VariablePresentationHint() { Kind = VariablePresentationHint.KindValue.Data },
                 Value = $"0x{stack[i]:X2}",
             });
